rename sudoku example
