+++ src/examples/sudoku/index.ts
@@ -1,4 +1,4 @@
-import { deploy, submitSolution, getZkappState } from './sudoku-snapp.js';
+import { deploy, submitSolution, getZkappState } from './sudoku-zkapp.js';
 import { cloneSudoku, generateSudoku, solveSudoku } from './sudoku-lib.js';
 import { shutdown } from 'snarkyjs';
